requirements.txt
# Core dependencies
pluggy>=1.6.0
ccxt>=4.0.0
pydantic>=2.0.0
typer>=0.9.0
promptantic>=1.1.0
prompt-toolkit>=3.0.48
textual>=6.11.0
cryptography>=41.0.0
pyyaml>=6.0.0
rich>=13.0.0
cachetools>=5.5.2
cachetools-async>=0.0.5
humanfriendly>=10.0
tenacity>=8.2.2

# Database
clickhouse-connect>=0.8.0  # 官方异步驱动，使用 HTTP 协议
# Technical Analysis
numpy>=1.24.0
pandas>=2.0.0
TA-Lib>=0.4.28
matplotlib>=3.7.0

# Point Process Modeling
# sparklen>=1.0.0  # do not support 3.14 yet

# Event System
pyee>=11.0.0

# Notifications
apprise>=1.9.0

# Statistical Analysis
hurst>=0.0.5

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# ruff>=0.1.0
